+++ packages/spacecat-shared-data-access/test/unit/service/index.test.js
@@ -109,10 +109,15 @@ describe('Data Access Object Tests', () => {
     'getApiKeyById',
   ];
 
+  const electroServiceFunctions = [
+    'Opportunity',
+    'Suggestion',
+  ];
+
   let dao;
 
   before(() => {
-    dao = createDataAccess();
+    dao = createDataAccess({ tableNameData: 'test' });
   });
 
   it('contains all known audit functions', () => {
@@ -175,7 +180,9 @@ describe('Data Access Object Tests', () => {
       ...importJobFunctions,
       ...importUrlFunctions,
       ...experimentFunctions,
-      ...apiKeyFunctions]);
+      ...apiKeyFunctions,
+      ...electroServiceFunctions,
+    ]);
     Object.keys(dao).forEach((funcName) => {
       expect(expectedFunctions).to.include(funcName);
     });