+++ packages/spacecat-shared-dynamo/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@adobe/spacecat-shared-dynamo-v1.3.49](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-dynamo-v1.3.48...@adobe/spacecat-shared-dynamo-v1.3.49) (2024-11-16)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#444](https://github.com/adobe/spacecat-shared/issues/444)) ([5e86c98](https://github.com/adobe/spacecat-shared/commit/5e86c98f274e514e8ec3032b52a128b7b394f5bc))
+
 # [@adobe/spacecat-shared-dynamo-v1.3.48](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-dynamo-v1.3.47...@adobe/spacecat-shared-dynamo-v1.3.48) (2024-11-11)